+++ packages/react/src/actions/evm/index.ts
@@ -1,0 +1,2 @@
+export * from './getEVMToken.js';
+export * from './getInjectedConnector.js';