+++ packages/react/src/chains/evm/testnets/unichainSepolia.ts
@@ -1,0 +1,28 @@
+import { EVMChain } from '../../../types/index.js';
+
+export const unichainSepolia: EVMChain = {
+  id: 1301,
+  name: 'Unichain Sepolia',
+  type: 'evm',
+  nativeCurrency: {
+    decimals: 18,
+    name: 'Ether',
+    symbol: 'ETH',
+  },
+  rpcUrls: {
+    default: {
+      http: ['https://unichain-sepolia-rpc.publicnode.com'],
+    },
+  },
+  blockExplorers: {
+    default: {
+      name: 'Unichain Sepolia Explorer',
+      url: 'https://unichain-sepolia.blockscout.com',
+    },
+  },
+  contracts: {
+    multicall3: {
+      address: '0xcA11bde05977b3631167028862bE2a173976CA11',
+    },
+  },
+};