+++ packages/react/src/providers/TronProvider.tsx
@@ -48,8 +48,10 @@ export const TronProvider = ({
       if (tronStore.getState().connectedAdapter?.name === this.name) setAddress(address);
       const connector = tronStore.getState().connectors[this.name];
 
+      console.log('[TRON] handleConnect', address, window.tron?.tronWeb);
+
       if (window.tron?.tronWeb) {
-        setTronWeb(window.tron.tronWeb);
+        if (window.tron.tronWeb) setTronWeb(window.tron.tronWeb);
         setConnector({ adapter: this, account: address, network: connector.network, readyState: this.state });
       }
     },
@@ -65,7 +67,7 @@ export const TronProvider = ({
       if (tronStore.getState().connectedAdapter?.name === this.name) setAddress(address);
       const connector = tronStore.getState().connectors[this.name];
       if (window.tron?.tronWeb) {
-        setTronWeb(window.tron.tronWeb);
+        if (window.tron.tronWeb) setTronWeb(window.tron.tronWeb);
         setConnector({ adapter: this, account: address, network: connector.network, readyState: this.state });
       }
     },
@@ -129,27 +131,31 @@ export const TronProvider = ({
   ]);
 
   const { mutateAsync: autoConnect } = useMutation({
-    mutationKey: ['tron autoConnect'],
+    mutationKey: ['autoConnectTron'],
     mutationFn: async () => {
       if (!connectedAdapter) {
         throw new Error('Connected adapter not found');
       }
       await connectedAdapter.connect();
       const tronWeb = window.tron?.tronWeb;
-      if (!tronWeb) {
-        throw new Error('TronWeb not found');
-      }
+
       return { account: connectedAdapter.address, chainId: undefined, adapter: connectedAdapter, tronWeb };
     },
     onSuccess: (data) => {
       setConnectedAdapter(data.adapter);
-      setTronWeb(data.tronWeb);
+      if (data.tronWeb) setTronWeb(data.tronWeb);
     },
   });
 
   // autoconnect
   useEffect(() => {
     if (!connectedAdapter || connectedAdapter.state !== AdapterState.Disconnect) {
+      if (connectedAdapter?.state === AdapterState.Connected) {
+        const tronWeb = window.tron?.tronWeb;
+        if (tronWeb) {
+          setTronWeb(tronWeb);
+        }
+      }
       return;
     }
     autoConnect();
@@ -168,15 +174,12 @@ export const TronProvider = ({
 
       await adapter.connect();
       const tronWeb = window.tron?.tronWeb;
-      if (!tronWeb) {
-        throw new Error('TronWeb not found');
-      }
 
       return { account: adapter.address, chainId: undefined, adapter, tronWeb };
     },
     onSuccess: (data) => {
       setConnectedAdapter(data.adapter);
-      setTronWeb(data.tronWeb);
+      if (data.tronWeb) setTronWeb(data.tronWeb);
     },
   });