+++ packages/react/src/utils/createChainConfigs.ts
@@ -19,11 +19,10 @@ const createChainConfigs = (
     ton: [],
   };
 
-  const defaultChains = getDefaultSupportedChains(testnet);
-
   if (chains) {
     supportedChains = overrideChainConfig(chains, overrides);
   } else {
+    const defaultChains = getDefaultSupportedChains(testnet);
     // Override with custom configs
     supportedChains = overrideChainConfig(defaultChains, overrides);
   }
@@ -50,8 +49,10 @@ const overrideChainConfig = (
   for (const chains of Object.values(chainsByType)) {
     for (const chain of chains) {
       const chainId = chain.id.toString() as ChainId;
+      const chainConfigData = CHAIN_DATA[chainId] || {};
       const chainData = {
-        ...CHAIN_DATA[chainId],
+        ...chainConfigData,
+        ...chain,
         ...overrides?.[CHAIN_NAME[chainId]],
       } as ChainData;