+++ packages/react/src/utils/getDefaultSupportedChains.ts
@@ -2,8 +2,11 @@ import { bitcoin } from '../chains/bitcoin.js';
 import * as cosmos from '../chains/cosmos/index.js';
 import * as evm from '../chains/evm/index.js';
 import { solana } from '../chains/solana.js';
+import { solanaTestnet } from '../chains/solana.testnet.js';
 import { sui } from '../chains/sui.js';
+import { suiTestnet } from '../chains/sui.testnet.js';
 import { tronMainnet } from '../chains/tron.js';
+import { tronShasta } from '../chains/tron.shasta.js';
 import {
   CosmsosChainType,
   EVMChain,
@@ -28,9 +31,30 @@ const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => 
 
   if (testnet) {
     // testnet chains
+    supportedChains.evm = [
+      evm.sepolia,
+      evm.holesky,
+      evm.arbitrumSepolia,
+      evm.avalancheFuji,
+      evm.amoy,
+      evm.abstractSepolia,
+      evm.beraChainTestnet,
+      evm.firechainTestnet,
+      evm.oasisSapphireTestnet,
+      evm.pentagonTestnet,
+      evm.soneiumTestnet,
+      evm.storyOdyssey,
+      evm.unichainSepolia,
+      evm.monadTestnet,
+    ] as EVMChain[];
+    supportedChains.cosmos = [];
+    supportedChains.solana = [solanaTestnet] as OtherChainData<'solana'>[];
+    supportedChains.tron = [tronShasta] as TronChain[];
+    supportedChains.sui = [suiTestnet] as SuiChainType[];
   } else {
     // mainnet chains
     supportedChains.evm = [
+      evm.abstract,
       evm.arbitrum,
       evm.arthera,
       evm.avalanche,
@@ -47,7 +71,9 @@ const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => 
       evm.matchain,
       evm.metis,
       evm.hyperliquid,
+      evm.ink,
       evm.mode,
+      evm.morph,
       evm.nahmii,
       evm.nero,
       evm.oasisSapphire,
@@ -63,11 +89,15 @@ const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => 
       evm.scroll,
       evm.shido,
       evm.soneium,
+      evm.sonic,
       evm.taiko,
       evm.tangle,
       evm.vanar,
-      evm.zkSync,
+      evm.worldChain,
       evm.xLayer,
+      evm.zero,
+      evm.zkSync,
+      evm.zora,
     ] as EVMChain[];
     supportedChains.cosmos = [cosmos.osmosis, cosmos.injective, cosmos.self, cosmos.router] as CosmsosChainType[];
     supportedChains.solana = [solana] as OtherChainData<'solana'>[];